Move recording stream to its own hpp/cpp
